Moved `Can` traits from internal to its own module
Closes https://github.com/PistonDevelopers/graphics/issues/661
